Add chat message parsers
Implement parsers and chat message parsers. A chat message will from now on be created by using a parser bound to the container. Implement a default parser that only splits the message by whitespace, but make the interface extendable for the user. Change up the chat message model and tests a bit.