Add roles
Add roles to players and commands. Extend the base command class to provide a method which returns a value indicating if the player has access to the command based on his roles. A player can invoke a command if the command does not specify any roles or if he has at least one role that the command specifies.